+++ src/templates/Authors.tsx
@@ -7,15 +7,16 @@ import { graphql } from 'gatsby';
 import { GatsbyImage, getImage } from 'gatsby-plugin-image';
 import styled from 'styled-components';
 import { useIntl } from 'react-intl';
+import { SchemedImage, Theme } from '@visiosto/components';
 
 import Intl from '../components/Intl';
 import Layout from '../components/layout/Layout';
 import LocalizedLink from '../components/link/LocalizedLink';
 import Rule from '../components/Rule';
-import SchemedImage from '../components/SchemedImage';
-import Theme from '../components/Theme';
 
 import createInternationalization from '../util/createInternationalization';
+
+import theme from '../theme';
 
 const Separator = styled.div`
   display: flex;
@@ -184,8 +185,8 @@ function Page({ data, pageContext }) {
                       >
                         <InstagramImage
                           alt={intl('footerInstagramImageText')}
-                          dark={getImage(data.instagram)}
-                          light={getImage(data.instagramColor)}
+                          dark={getImage(data.instagram)!}
+                          light={getImage(data.instagramColor)!}
                         />
                       </a>
                     );
@@ -203,8 +204,8 @@ function Page({ data, pageContext }) {
                       >
                         <SocialMediaImage
                           alt={intl('footerFacebookImageText')}
-                          dark={getImage(data.facebook)}
-                          light={getImage(data.facebookColor)}
+                          dark={getImage(data.facebook)!}
+                          light={getImage(data.facebookColor)!}
                         />
                       </a>
                     );
@@ -222,8 +223,8 @@ function Page({ data, pageContext }) {
                       >
                         <TwitterImage
                           alt={intl('footerTwitterImageText')}
-                          dark={getImage(data.twitter)}
-                          light={getImage(data.twitterColor)}
+                          dark={getImage(data.twitter)!}
+                          light={getImage(data.twitterColor)!}
                         />
                       </a>
                     );
@@ -241,8 +242,8 @@ function Page({ data, pageContext }) {
                       >
                         <LinkedinImage
                           alt={intl('footerLinkedinImageText')}
-                          dark={getImage(data.linkedin)}
-                          light={getImage(data.linkedinColor)}
+                          dark={getImage(data.linkedin)!}
+                          light={getImage(data.linkedinColor)!}
                         />
                       </a>
                     );
@@ -260,8 +261,8 @@ function Page({ data, pageContext }) {
                       >
                         <GithubImage
                           alt={intl('footerGithubImageText')}
-                          dark={getImage(data.github)}
-                          light={getImage(data.github)}
+                          dark={getImage(data.github)!}
+                          light={getImage(data.github)!}
                         />
                       </a>
                     );
@@ -288,7 +289,7 @@ function Authors({ data, pageContext }) {
   const { locale } = pageContext;
   return (
     <Intl locale={simpleLocales[locale.replace('-', '_')]}>
-      <Theme>
+      <Theme theme={theme}>
         <Page data={data} pageContext={pageContext} />
       </Theme>
     </Intl>