+++ src/templates/Index.tsx
@@ -12,6 +12,7 @@ import {
   PaperAirplaneIcon,
   PencilIcon,
 } from '@primer/octicons-react';
+import { Theme } from '@visiosto/components';
 
 import AuthorContactCard from '../components/AuthorContactCard';
 import Break from '../components/Break';
@@ -20,7 +21,8 @@ import Cover from '../components/Cover';
 import FeatureCard from '../components/FeatureCard';
 import Intl from '../components/Intl';
 import LayoutIndex from '../components/layout/LayoutIndex';
-import Theme from '../components/Theme';
+
+import theme from '../theme';
 
 const H2 = styled.h2`
   font-size: 2.2rem;
@@ -169,7 +171,7 @@ function Index({ data, pageContext }) {
   const { locale } = pageContext;
   return (
     <Intl locale={simpleLocales[locale.replace('-', '_')]}>
-      <Theme>
+      <Theme theme={theme}>
         <Page data={data} pageContext={pageContext} />
       </Theme>
     </Intl>